ML-Horizons-Horace-backend/requirements.txt
pymongo
langchain
setuptools
wheel
numpy
langchain_community
python-dotenv
langgraph
langchain-ollama
langsmith
langchain_mongodb
boto3
botocore
python-dotenv
scrapingbee
bs4
flask
langchain_cohere
langchain_aws